rename : 여행 기록 package 이름 변경
