+++ src/main/java/com/crude/travelcrew/domain/record/model/entity/RecordComment.java
@@ -1,0 +1,47 @@
+package com.crude.travelcrew.domain.record.model.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+
+import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.global.entity.BaseTime;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class RecordComment extends BaseTime {
+
+	@Id
+	@Column(name = "record_comment_id")
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column(nullable = false)
+	private String content;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "member_id", updatable = false)
+	private Member member;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "record_id", updatable = false)
+	private Record record;
+
+	public void update (String content) {
+		this.content = content;
+	}
+}
