+++ src/main/java/com/crude/travelcrew/domain/record/repository/RecordHeartRepository.java
@@ -1,0 +1,16 @@
+package com.crude.travelcrew.domain.record.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.domain.record.model.entity.Record;
+import com.crude.travelcrew.domain.record.model.entity.RecordHeart;
+
+public interface RecordHeartRepository extends JpaRepository<RecordHeart, Long> {
+
+	boolean existsByRecordAndMember(Record record, Member member);
+
+	Optional<RecordHeart> findByRecordAndMember(Record record, Member member);
+}
