+++ src/main/java/com/crude/travelcrew/domain/record/repository/custom/CustomRecordCommentRepository.java
@@ -1,0 +1,6 @@
+package com.crude.travelcrew.domain.record.repository.custom;
+
+public interface CustomRecordCommentRepository {
+
+	Long deleteAllByRecordId(Long recordId);
+}
