+++ src/main/java/com/crude/travelcrew/domain/record/repository/custom/CustomRecordImageRepository.java
@@ -1,0 +1,6 @@
+package com.crude.travelcrew.domain.record.repository.custom;
+
+public interface CustomRecordImageRepository {
+
+	Long deleteAllByRecordId(Long recordId);
+}
