--- src/main/java/com/crude/travelcrew/domain/travelrecord/model/dto/EditRecordReq.java
@@ -1,21 +1,0 @@
-package com.crude.travelcrew.domain.travelrecord.model.dto;
-
-import javax.validation.constraints.NotBlank;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class EditRecordReq {
-
-	@NotBlank(message = "제목을 입력해주세요.")
-	private String title;
-
-	@NotBlank(message = "내용을 입력해주세요.")
-	private String content;
-}
