--- src/main/java/com/crude/travelcrew/domain/travelrecord/repository/RecordRepository.java
@@ -1,8 +1,0 @@
-package com.crude.travelcrew.domain.travelrecord.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.crude.travelcrew.domain.travelrecord.model.entity.Record;
-
-public interface RecordRepository extends JpaRepository<Record, Long> {
-}
