--- src/main/java/com/crude/travelcrew/domain/travelrecord/repository/custom/CustomRecordCommentRepository.java
@@ -1,6 +1,0 @@
-package com.crude.travelcrew.domain.travelrecord.repository.custom;
-
-public interface CustomRecordCommentRepository {
-
-	Long deleteAllByRecordId(Long recordId);
-}
