--- src/main/java/com/crude/travelcrew/domain/travelrecord/repository/custom/CustomRecordImageRepository.java
@@ -1,6 +1,0 @@
-package com.crude.travelcrew.domain.travelrecord.repository.custom;
-
-public interface CustomRecordImageRepository {
-
-	Long deleteAllByRecordId(Long recordId);
-}
